Preliminary version of an algorithm that expands our list of machines into a complete list of all machines categorizing them by result (halt, infinite, unknown).
As of now it attempts to count the number in each category, although it is
not working correctly yet (or our list of machines is incorrect!)

It also includes a very cool enumerator and unenumerator which turn a
transition table (with no blank cells) into an integer and vice-versa
(currently incomplete because unenumerator needs to be given a number of
states and symbols).  If we find wider use for this algorithm it will be
given it's own file.