Very, very basic script for inputing machines into Computer readable format.
makes a file based on the num_states, num_symbols and name input in the
Machines directory and then asks for each cell's specification.  cell
input is very basic and only accepts symbol-dir-state order with
limited format.

Should be updated to accept multiple formats that others used for
presentation (hopefully autorecognize) and such.