Yay for abstraction.
The Numbers folder is being used to store the abstraction for the
numberical classes that I'm writing for algebraic manipulation and
fractions.

The Macro folder is being used for abstractions to macro machines and
accelerated simulation.  NEW! now supports backsymbols.

Quick_Sim.py is just a quick hack to use the tools in Macro/

usage: Quick_Sim.py macro_size filename

NOTE:
With backsymbols, I am able to simulate Marxen's Giant 6x2 machine to
completion in ~24 seconds!!! (macro_size=2, I think)

Also Note:
While Quick_Sim is pretty cool, it only touches the surface of what we
can do with this new abstraction.  It lays a backsymbol macro machine on
a block macro machine on top of a normal TM.  But with the new
abstraction we can put any number of layers of macro-machines atop
macro-machines (If properly implimented this could speed up counter and
counter-like computation a lot!)