My new CTL implimentation that takes care of the trivial regular expression.
The filter will take the following parameters:
cutoff=200 The number of steps to run in the TM before trying the CTL
size=1     The block size for the mactro machine
offset=0   The offset for the macro machine (so that the blocks align
with the actual machine)

experimentally cutoff needs to be at least 50 and beyond 100 doesn't
make much difference (specifically tested on 2x4's).  I chose 200 as a
safety, 1000 ought to be plenty for most applications.

size will catch different machines just like in the tree-filter

offset shifts the blocks that the macro machine uses for machines that
would be caught by a macro machine but are offset.

range: 0 <= offset < size


You can also run Simple_CTL.py directly to get detail on the CTL
attempt for a single machine.

Syntax: Simple_CTL.py filename line_num cutoff block_size offset