Added at "timeout" option.  If specified, this sets a timer so that Chain_Simulator (and Turing_Machine) will return unsuccessfully when the timer goes off.
If "timeout" is not specified then the behavior is as before.  Similarly,
if the "set_time_out()" method for "Chain_Simulator" is not called, then
the behavior of everything in "Chain_Simulator" is unchanged - promise,
;-).

Finally, this code may not be catching everywhere it would be useful to
have this type of "timeout" feature but it's a start!