Counts nuber of TM represented by the descriptions in a file.  For example we force Tree Normal Form and don't define a transition until necessary, thus many machines are rep by each of ours.
I originally wrote this for fun.  But I recently noticed that it could
be used to make sure that we have all n x m TMs (We haven't lost any
along the way) and that Assign_Undecided doesn't gain or lose any
machines.

You see the number of machines represented doesn't change when we
assign the undefined transitions (even if the number of reps
increases!), so a before and after count should be the same.

Also it can easily be shown that the total count for all n-state
m-symbol machines is:

(n*m - 1) * (2*n*m)^(n*m - 2)

[With the restriction that A1->1RB and there is a single halting
transition 1RH.  Which is the restriction used by Count.py]

Thus at any time running Count on all n x m files should get this value.