Parallel Tempering! Turns out I wrote this while I was taking the class.
Real simple:
PT.PT(B, neigh, f, C, steps)
Runs Parallel Tempering with temperatures [1/b for b in B] in order to
maximize utility function f. neigh(s) returns a random neighboring state
to s. It starts with configurations C[i] and runs for steps in parallel
returning the final configurations.

Currently it does not save the best configuration (TODO).

It copies the configuration each time it chooses a neighbor (instead of
editting in place, this could get expensive, but not with the BB
probably.

-Shawn