Changed to allow TM output to be compressed as it is written using "iostreams" from Boost.
"cmake" detects if Boost is available. If not, the code is compiled without
the option compress TM output and will generate an error if compression
is requested.

If Boost is available, the code is compiled to use it but the user still
controls if TM output is compressed. By default, compression isn't used...